requirements.txt
GetOldTweets3==0.0.11
python-dotenv==0.13.0
tweepy==3.8.0
